refactor(glossary): streamline glossary handling and introduce Crowdin integration
